Move depalettize-shader generator to Common
